+++ src/components/general/Timer/timer.module.scss
