+++ src/components/general/Wrapper/wrapper.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .wrapper {
   width: $wrapper-size;
