+++ src/components/header/index.tsx
@@ -2,8 +2,8 @@ import { FC, useContext, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { ISprint } from '../../types/sprint';
 import style from './header.module.scss';
-import Timer from '../common/Timer';
-import Wrapper from '../common/Wrapper';
+import Timer from '../general/Timer';
+import Wrapper from '../general/Wrapper';
 import LanguageContext from '../../contexts/LanguageContext';
 import ruIcon from '../../assets/icons/ru.png';
 import enIcon from '../../assets/icons/en.png';
