+++ src/components/modal/AdditionSection/additionSection.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/mixins.scss';
-@import '../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
+@import '../../../styles/variables.scss';
 
 .additionSection {
   @include flex-column;
