+++ src/components/modal/Comments/comments.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables.scss';
-@import '../../styles/mixins.scss';
+@import '../../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
 
 .commentSection {
   @include flex-column;
