+++ src/components/modal/MainSection/mainSection.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/mixins.scss';
-@import '../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
+@import '../../../styles/variables.scss';
 
 .mainSection {
   @include flex-column;
