+++ src/components/modal/StatusSelector/statusSelector.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .selection {
   margin-bottom: $bottom-margin;
