+++ src/components/modal/TaskDetail/taskDetail.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .detail {
   display: flex;
