+++ src/components/table/Column/column.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/mixins.scss';
-@import '../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
+@import '../../../styles/variables.scss';
 
 .column {
   @include flex-column;
