+++ src/components/table/Participants/participants.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .participants {
   display: flex;
