+++ src/components/table/TaskCard/taskCard.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables.scss';
-@import '../../styles/mixins.scss';
+@import '../../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
 
 .card {
   margin: $main-margin;
