+++ src/components/table/TasksTable/tasksTable.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .table {
   width: 100%;
